fix(start): fail loud on runtime.start() failure + split _start.py
Operator directive (fail loud, no silent feedback): agent_start raised a
cause-less RuntimeError when runtime.start() returned False — the real boot
failure (boot-drain timeout, auth, broken in-container shell, immediate claude
exit) was swallowed, no log, no diagnostic. Found via a neutral smoke agent.

- agent_start now captures the agent's tmux pane (TmuxManager.capture_logs) +
  session_exists and raises a descriptive RuntimeError instead of a bare one.
- Split _start.py (544 > 512-line limit) — extracted the four pre-flight
  helpers (_verify_real_liveness, _resolve_strict_drift,
  _rotate_to_healthy_account, _check_spec_source_drift_at_launch) into
  _start_preflight.py; _start.py re-exports them for back-compat.

Verified: 456 _lifecycle tests collect clean; test__start_drift + test__start_spawn_acl pass (12).
